bug fixes to test

+++ src/utils/cloudinary.js
@@ -1,26 +1,27 @@
-import {v2 as cloudinary} from 'cloudinary';
-import fs from 'fs';
+import { v2 as cloudinary } from "cloudinary";
+import fs from "fs";
 
-
-cloudinary.config({ 
-    cloud_name:process.env.CLOUDINARY_CLOUD_NAME, 
-    api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret:process.env.CLOUDINARY_API_SECRET  // Click 'View API Keys' above to copy your API secret
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-
-const uploadOnCloudinary=async(localFilePath)=>{
-   try {
-    const response= cloudinary.uploader.upload(localFilePath,{
-        resource_type:auto
+const uploadOnCloudinary = async (localFilePath) => {
+  try {
+    if (!localFilePath) return null;
+    //upload the file on cloudinary
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto",
     });
-
-    console.log("File uploaded successfully on cloudinary",response.url);
-
-   } catch (error) {
+    // file has been uploaded successfull
+    //console.log("file is uploaded on cloudinary ", response.url);
     fs.unlinkSync(localFilePath);
+    return response;
+  } catch (error) {
+    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
     return null;
-   }
-}
+  }
+};
 
-export {uploadOnCloudinary};
+export { uploadOnCloudinary };
